+++ examples/setup_standard_tables.py
@@ -29,5 +29,6 @@ def main():
         })
     dwg.saveas('standard_linetypes.dxf')
 
+
 if __name__ == '__main__':
     main()