+++ tests/test_complex_line_type.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ezdxf.tools.complex_ltype import lin_tokenizer, lin_parser, lin_compiler
+from ezdxf.tools.complex_ltype import lin_tokenizer, lin_parser, lin_compiler, ComplexLineTypePart
 
 
 def test_line_type_tokenizer_just_numbers():
@@ -76,10 +76,45 @@ def test_lin_compiler_shape():
     assert text.type == 'SHAPE'
     assert text.text == 'BOX'
     assert text.font == 'ltypeshp.shx'
-    assert text.tags[0] == (44, -.1)
-    assert text.tags[1] == (46, .1)
+    assert text.tags[0] == (46, .1)
+    assert text.tags[1] == (50, 0.)
+    assert text.tags[2] == (44, -.1)
+    assert text.tags[3] == (45, 0.)
     assert result[3] == (49, -.1)
     assert result[4] == (49, 1)
+
+
+def test_tags_from_complex_text():
+    ltype = 'A,.5,-.2,["GAS",STANDARD,S=.1,U=0.0,X=-0.1,Y=-.05],-.25'
+    result = lin_compiler(ltype)
+    cx_part = result[2]
+    assert isinstance(cx_part, ComplexLineTypePart)
+    assert cx_part.type == 'TEXT'
+    tags = cx_part.complex_ltype_tags(None)
+    assert tags[0] == (74, 2)
+    assert tags[1] == (75, 0)
+    assert tags[2] == (340, 0)  # default handle with a drawing
+    assert tags[3] == (46, .1)     # s
+    assert tags[4] == (50, 0)      # r
+    assert tags[5] == (44, -0.1)   # x
+    assert tags[6] == (45, -0.05)  # y
+    assert tags[7] == (9, 'GAS')
+
+
+def test_tags_from_complex_shape():
+    ltype = 'A,.25,-.1,[BOX,ltypeshp.shx,x=-.1,s=.1],-.1,1'
+    result = lin_compiler(ltype)
+    cx_part = result[2]
+    assert isinstance(cx_part, ComplexLineTypePart)
+    assert cx_part.type == 'SHAPE'
+    tags = cx_part.complex_ltype_tags(None, shapes_table={'BOX': 17})
+    assert tags[0] == (74, 4)
+    assert tags[1] == (75, 17)   # as in the shapes table defined
+    assert tags[2] == (340, 0)   # default handle with a drawing
+    assert tags[3] == (46, .1)   # s
+    assert tags[4] == (50, 0.)   # r
+    assert tags[5] == (44, -.1)  # x
+    assert tags[6] == (45, 0.)   # y
 
 
 if __name__ == '__main__':